requirements.txt
asgiref==3.2.3
astroid==2.3.3
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Django==3.0.2
google-api-python-client==1.7.11
google-auth==1.10.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
gunicorn==20.0.4
httplib2==0.16.0
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.4.4
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.14.0
sqlparse==0.3.0
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.7
wrapt==1.11.2
